Feature/rental (#12)
* add rent

* add images

* add gem omniauth

* add auth with github

* add mailer for rental

* add sort for rental

* seed update

* add ru language

* add style

* refactor code

* add pundit

* fix style

* fix link

+++ app/controllers/products_controller.rb
@@ -1,29 +1,32 @@
 # frozen_string_literal: true
 
 class ProductsController < ApplicationController
-  NUMBER_ITEMS_PER_PAGE = 19
+  NUMBER_ITEMS_PER_PAGE = 4
+
   def index
-    @products = Product.all.page(params[:page]).per(NUMBER_ITEMS_PER_PAGE)
+    @products = Product.includes(:salon).includes(:product_type).includes(:avatar_attachment).all.page(params[:page]).per(NUMBER_ITEMS_PER_PAGE)
   end
 
   def menu
     @list_salons = current_user.profileable.salons
-    @products = Product.where(salon_id: @list_salons)
+    @products = Product.includes(:avatar_attachment).where(salon_id: @list_salons)
   end
 
   def show
     @product = Product.find(params[:id])
+    authorize @product
   end
 
   def new
     @product = Product.new
+    authorize @product
   end
 
   def create
     @product = Product.create(product_params)
     if @product
       flash[:success] = 'Success'
-      redirect_to product_menu_path(@product)
+      redirect_to product_menu_path
     else
       flash[:error] = 'Error'
       render :new
@@ -32,6 +35,7 @@ class ProductsController < ApplicationController
 
   def edit
     @product = Product.find(params[:id])
+    authorize @product
   end
 
   def update
@@ -52,7 +56,7 @@ class ProductsController < ApplicationController
 
   def destroy
     @product = Product.find(params[:id])
-
+    authorize @product
     if @product.destroy
       flash[:success] = 'Success'
       if current_user[:role] == 'Admin'
@@ -68,6 +72,6 @@ class ProductsController < ApplicationController
   private
 
   def product_params
-    params.require(:product).permit(:name, :description, :price, :product_type_id, :salon_id)
+    params.require(:product).permit(:name, :description, :price, :product_type_id, :salon_id, :avatar)
   end
 end
